+++ src/components/UI/Snackbar/index.tsx
@@ -1,5 +1,5 @@
 // deps
-import React, { useContext } from 'react'
+import React from 'react'
 import Snackbar from '@material-ui/core/Snackbar'
 import Alert from '@material-ui/lab/Alert'
 // model
@@ -11,7 +11,7 @@ const SnackbarComp: React.FC = () => {
   const {
     snackbarState: { message, open, type },
     setSnackbarState,
-  } = useContext<ISnackbarContext>(SnackbarContext)
+  } = React.useContext<ISnackbarContext>(SnackbarContext)
 
   const handleClose = () => {
     setSnackbarState({ open: false })
@@ -26,6 +26,7 @@ const SnackbarComp: React.FC = () => {
       }}
       open={open}
       autoHideDuration={6000}
+      onClose={handleClose}
     >
       <Alert onClose={handleClose} severity={type}>
         {message}
