--- src/containers/Employee/hooks/useSubmitEmployeeModal.ts
@@ -1,11 +1,0 @@
-// deps
-import React from 'react'
-// model
-import { IEmployeeModalProps } from '_/containers/Employee/components/DetailsModal/types'
-// helpers
-
-// export const useSubmitEmployeeModal = async (props: IEmployeeModalProps) => {
-export const useSubmitEmployeeModal = () => {
-  const test = () => {}
-  return [test]
-}
