--- src/containers/Employees/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useGetEmployees } from './useGetEmployees'
