+++ src/context/directories.ts
@@ -1,0 +1,7 @@
+// deps
+import React from 'react'
+// components
+// model
+// helpers
+
+const DirectoriesContext = React.createContext({})
